V2.0.0
增加GNU/GCC工具链支持
支持CMake构建系统
优化Keil/GCC双工具链架构

+++ ble/prf/prf_bass.c
@@ -21,7 +21,7 @@
 
 #if (DBG_BASS)
 #include "dbg.h"
-#define DEBUG(format, ...)    debug("<%s,%d>" format "\r\n", __MODULE__, __LINE__, ##__VA_ARGS__)
+#define DEBUG(format, ...)    debug("<%s,%d>" format "\r\n", __MODULE__, (int)__LINE__, ##__VA_ARGS__)
 #else
 #define DEBUG(format, ...)
 #define debugHex(dat,len)
