+++ projects/bleHid/src/proc.c
@@ -19,7 +19,7 @@
 
 #if (DBG_PROC)
 #include "dbg.h"
-#define DEBUG(format, ...)    debug("<%s,%d>" format "\r\n", __MODULE__, __LINE__, ##__VA_ARGS__)
+#define DEBUG(format, ...)    debug("<%s,%d>" format "\r\n", __MODULE__, (int)__LINE__, ##__VA_ARGS__)
 #else
 #define DEBUG(format, ...)
 #define debugHex(dat, len)
@@ -36,16 +36,16 @@ enum uart_cmd
     CMD_KB_REP        = 0xC0,
     CMD_MEDIA_REP,
     CMD_SYSTEM_REP,
-    
+
     CMD_MOUSE_REP     = 0xD0,
     CMD_MOUSE_BTN,
     CMD_MOUSE_X,
     CMD_MOUSE_Y,
     CMD_MOUSE_WHEEL,
-    
+
     CMD_CONN_INTERVAL = 0xE0,
     CMD_SVC_CHG,
-    
+
     CMD_BATT_LVL      = 0xF0,
     CMD_BATT_PWR_STA
 };
@@ -70,7 +70,7 @@ static void uart_proc(void)
     static uint8_t null_cnt = 0;
     uint16_t len;
     bool finish = true;
-    
+
     len = uart1Rb_Read(&buff[buff_len], CMD_MAX_LEN - buff_len);
     if (len > 0)
     {
@@ -96,7 +96,7 @@ static void uart_proc(void)
     if (app_state_get() == APP_CONNECTED)
     {
         switch (buff[0])
-        {    
+        {
             case CMD_KB_REP:
             {
                 DEBUG("key_code:0x%02x", buff[1]);
@@ -104,7 +104,7 @@ static void uart_proc(void)
                 keybd_report_send(app_env.curidx, kyebd_report);
                 keybd_report_send(app_env.curidx, NULL);
             } break;
-            
+
             case CMD_MEDIA_REP:
             {
                 if (buff_len >= 4)
@@ -115,11 +115,11 @@ static void uart_proc(void)
                     media_report_send(app_env.curidx, NULL);
                 }
             } break;
-            
+
             case CMD_SYSTEM_REP:
             {
                 DEBUG("CMD_SYSTEM_REP:%x", buff[1]);
-               
+
                 system_report_send(app_env.curidx, &buff[1]);
                 system_report_send(app_env.curidx, NULL);
             } break;
@@ -130,6 +130,7 @@ static void uart_proc(void)
                 buff[0] = CMD_MOUSE_X;
                 buff[1] = 10;
             } //no break;
+            /*fallthrough*/
             case CMD_MOUSE_BTN:
             case CMD_MOUSE_X:
             case CMD_MOUSE_Y:
@@ -138,10 +139,10 @@ static void uart_proc(void)
                 uint8_t mouse_data[] = {0, 0, 0, 0, 0, 0};
                 uint8_t idx = buff[0]-CMD_MOUSE_BTN;
                 if (idx) idx = (idx*2) - 1; // 2X 2Y
-                
+
                 mouse_data[idx] = buff[1];
                 DEBUG("MOUSE(1Btn,2X,2Y,1Wheel)[%d]:0x%02x", idx, buff[1]);
-                
+
                 mouse_report_send(app_env.curidx, mouse_data);
                 mouse_report_send(app_env.curidx, NULL); // release
             } break;
@@ -151,17 +152,17 @@ static void uart_proc(void)
                 DEBUG("Batt Lvl");
                 bass_bat_lvl_update(buff[1]);
             } break;
-            
+
             case CMD_BATT_PWR_STA:
             {
                 DEBUG("Batt Power State");
                 bass_pwr_sta_update(buff[1]);
             } break;
-            
+
             case CMD_CONN_INTERVAL:
             {
                 struct gapc_conn_param conn_pref =
-                { 
+                {
                     /// Connection interval minimum unit in 1.25ms
                     .intv_min = 6,
                     /// Connection interval maximum unit in 1.25ms
@@ -171,7 +172,7 @@ static void uart_proc(void)
                     /// Connection supervision timeout multiplier unit in 10ms
                     .time_out = 300,
                 };
-                
+
                 conn_pref.intv_min = buff[1];
                 conn_pref.intv_max = buff[2];
                 conn_pref.latency  = buff[3];
@@ -179,13 +180,13 @@ static void uart_proc(void)
                 DEBUG("intvn:%d, intvm:%d, lat:%d", conn_pref.intv_min, conn_pref.intv_max, conn_pref.latency);
                 gapc_update_param(app_env.curidx, &conn_pref);
             } break;
-            
+
             case CMD_SVC_CHG:
             {
                 DEBUG("service change");
                 //gatt_svc_chg(app_env.curidx, 0, 0xFFFF);
             } break;
-            
+
             default:
             {
             } break;
