+++ projects/bleHid_Keybd/src/proc.c
@@ -20,7 +20,7 @@
 
 #if (DBG_PROC)
 #include "dbg.h"
-#define DEBUG(format, ...)    debug("<%s,%d>" format "\r\n", __MODULE__, __LINE__, ##__VA_ARGS__)
+#define DEBUG(format, ...)    debug("<%s,%d>" format "\r\n", __MODULE__, (int)__LINE__, ##__VA_ARGS__)
 #else
 #define DEBUG(format, ...)
 #define debugHex(dat, len)
@@ -38,9 +38,9 @@ enum uart_cmd
     CMD_KB_REP        = 0xC0,
     CMD_MEDIA_REP,
     CMD_SYSTEM_REP,
-    
+
     CMD_CONN_INTERVAL = 0xE0,
-    
+
     CMD_BATT_LVL      = 0xF0,
     CMD_BATT_PWR_STA
 };
@@ -62,7 +62,7 @@ static void uart_proc(void)
     static uint8_t null_cnt = 0;
     uint16_t len;
     bool finish = true;
-    
+
     len = uart1Rb_Read(&buff[buff_len], CMD_MAX_LEN - buff_len);
     if (len > 0)
     {
@@ -88,7 +88,7 @@ static void uart_proc(void)
     if (app_state_get() == APP_CONNECTED)
     {
         switch (buff[0])
-        {    
+        {
             case CMD_KB_REP:
             {
                 DEBUG("key_code:0x%02x", buff[1]);
@@ -96,7 +96,7 @@ static void uart_proc(void)
                 keybd_report_send(app_env.curidx, kyebd_report);
                 keybd_report_send(app_env.curidx, NULL);
             } break;
-            
+
             case CMD_MEDIA_REP:
             {
                 if (buff_len >= 4)
@@ -107,11 +107,11 @@ static void uart_proc(void)
                     media_report_send(app_env.curidx, NULL);
                 }
             } break;
-            
+
             case CMD_SYSTEM_REP:
             {
                 DEBUG("CMD_SYSTEM_REP:%x", buff[1]);
-               
+
                 system_report_send(app_env.curidx, &buff[1]);
                 system_report_send(app_env.curidx, NULL);
             } break;
@@ -121,17 +121,17 @@ static void uart_proc(void)
                 DEBUG("Batt Lvl");
                 bass_bat_lvl_update(buff[1]);
             } break;
-            
+
             case CMD_BATT_PWR_STA:
             {
                 DEBUG("Batt Power State");
                 bass_pwr_sta_update(buff[1]);
             } break;
-            
+
             case CMD_CONN_INTERVAL:
             {
                 struct gapc_conn_param conn_pref =
-                { 
+                {
                     /// Connection interval minimum unit in 1.25ms
                     .intv_min = 6,
                     /// Connection interval maximum unit in 1.25ms
@@ -141,7 +141,7 @@ static void uart_proc(void)
                     /// Connection supervision timeout multiplier unit in 10ms
                     .time_out = 300,
                 };
-                
+
                 conn_pref.intv_min = buff[1];
                 conn_pref.intv_max = buff[2];
                 conn_pref.latency  = buff[3];
@@ -149,7 +149,7 @@ static void uart_proc(void)
                 DEBUG("intvn:%d, intvm:%d, lat:%d", conn_pref.intv_min, conn_pref.intv_max, conn_pref.latency);
                 gapc_update_param(app_env.curidx, &conn_pref);
             } break;
-            
+
             default:
             {
             } break;
@@ -168,6 +168,6 @@ void user_procedure(void)
     #if (UART_CMD)
     uart_proc();
     #endif //(UART_CMD)
-    
+
     keys_scan();
 }
